# docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: Invalid credentials or user not in allowed group
        type: string
    type: object
...
  models.RuntimeImageActionResponse:
    properties:
      image:
        type: string
      message:
        type: string
      output:
        type: string
      success:
        type: boolean
    type: object
  models.RuntimeImagePullRequest:
    properties:
      image:
        type: string
    required:
    - image
    type: object
  models.RuntimeImageSummary:
    properties:
      createdAt:
        type: string
      id:
        type: string
      repoDigests:
        items:
          type: string
        type: array
      repoTags:
        items:
          type: string
        type: array
      shortId:
        type: string
      size:
        type: string
      virtualSize:
        type: string
    type: object
  models.RuntimeImagesResponse:
    properties:
      images:
        items:
          $ref: '#/definitions/models.RuntimeImageSummary'
        type: array
      runtime:
        type: string
    type: object
...
info:
  contact:
    name: Containerlab API Server maintainers
    url: https://github.com/srl-labs/clab-api-server/issues
  description: |-
    REST API for authenticated Linux users to manage Containerlab deployments, topology files, node access, captures, and host networking tools.

    Most endpoints under /api/v1 require a JWT from POST /login. Send it as: Authorization: Bearer <token>.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Containerlab API Server
  version: "1.0"
paths:
...
  /api/v1/images:
    delete:
      parameters:
      - description: Image reference or ID
        in: query
        name: reference
        required: true
        type: string
      - description: Force image removal
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RuntimeImageActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a runtime image
      tags:
      - Images
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RuntimeImagesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List runtime images
      tags:
      - Images
  /api/v1/images/pull:
    post:
      consumes:
      - application/json
      parameters:
      - description: Image reference
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RuntimeImagePullRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RuntimeImageActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Pull a runtime image
      tags:
      - Images
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"